Refactor some of the filesystem stuff
