+++ docs/contracts/filesystem/index.md
@@ -1,0 +1,9 @@
+---
+title: Filesystem Contract
+---
+
+## Installation
+
+```shell
+composer require sonsofphp/filesystem-contract
+```
