+++ src/SonsOfPHP/Component/Filesystem/Exception/FileNotFoundException.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace SonsOfPHP\Component\Filesystem\Exception;
 
+use SonsOfPHP\Contract\Filesystem\Exception\FileNotFoundExceptionInterface;
+
 /**
  * @author Joshua Estes <joshua@sonsofphp.com>
  */
-final class FileNotFoundException extends FilesystemException {}
+final class FileNotFoundException extends FilesystemException implements FileNotFoundExceptionInterface {}
