+++ src/SonsOfPHP/Component/Filesystem/Exception/FilesystemException.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace SonsOfPHP\Component\Filesystem\Exception;
 
 use Exception;
+use SonsOfPHP\Contract\Filesystem\Exception\FilesystemExceptionInterface;
 
 /**
  * @author Joshua Estes <joshua@sonsofphp.com>
  */
-class FilesystemException extends Exception {}
+class FilesystemException extends Exception implements FilesystemExceptionInterface {}
