+++ src/SonsOfPHP/Contract/Filesystem/Adapter/AdapterInterface.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace SonsOfPHP\Component\Filesystem\Adapter;
+namespace SonsOfPHP\Contract\Filesystem\Adapter;
 
-use SonsOfPHP\Component\Filesystem\ContextInterface;
-use SonsOfPHP\Component\Filesystem\Exception\FileNotFoundException;
-use SonsOfPHP\Component\Filesystem\Exception\FilesystemException;
+use SonsOfPHP\Contract\Filesystem\Exception\FileNotFoundExceptionInterface;
+use SonsOfPHP\Contract\Filesystem\Exception\FilesystemExceptionInterface;
+use SonsOfPHP\Contract\Filesystem\ContextInterface;
 
 /**
  * Base Adapter Interface that all adapters implement.
@@ -22,16 +22,16 @@ interface AdapterInterface
      *   Can either be a string or a resource. If not one of these types, the
      *   adapter should throw an exception
      *
-     * @throws FilesystemException When $contents is invalid argument
-     * @throws FilesystemException Generic Failure Exception
+     * @throws FilesystemExceptionInterface When $contents is invalid argument
+     * @throws FilesystemExceptionInterface Generic Failure Exception
      */
     public function add(string $path, mixed $contents, ?ContextInterface $context = null): void;
 
     /**
      * Returns the content of a given file found at $path
      *
-     * @throws FileNotFoundException When $path does not exist
-     * @throws FilesystemException   Generic Failure Exception
+     * @throws FileNotFoundExceptionInterface When $path does not exist
+     * @throws FilesystemExceptionInterface   Generic Failure Exception
      *
      * @return string|resouce
      *   This should return either a string or resouce
@@ -41,20 +41,20 @@ interface AdapterInterface
     /**
      * Deletes files and directories
      *
-     * @throws FileNotFouneException When $path does not exist
-     * @throws FilesystemException   Generic Failure Exception
+     * @throws FileNotFouneExceptionInterface When $path does not exist
+     * @throws FilesystemExceptionInterface   Generic Failure Exception
      */
     public function remove(string $path, ?ContextInterface $context = null): void;
 
     /**
      * Checks to see if a file or directory exists
      *
-     * @throws FilesystemException Generic Failure Exception
+     * @throws FilesystemExceptionInterface Generic Failure Exception
      */
     public function has(string $path, ?ContextInterface $context = null): bool;
 
     /**
-     * @throws FilesystemException Generic Failure Exception
+     * @throws FilesystemExceptionInterface Generic Failure Exception
      */
     public function isFile(string $path, ?ContextInterface $context = null): bool;
 }
