+++ src/SonsOfPHP/Contract/Filesystem/Exception/FileNotFoundExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Contract\Filesystem\Exception;
+
+/**
+ * @author Joshua Estes <joshua@sonsofphp.com>
+ */
+interface FileNotFoundExceptionInterface extends FilesystemExceptionInterface {}
