+++ src/SonsOfPHP/Contract/Filesystem/composer.json
@@ -21,7 +21,7 @@
     "support": {
         "issues": "https://github.com/SonsOfPHP/sonsofphp/issues",
         "forum": "https://github.com/orgs/SonsOfPHP/discussions",
-        "docs": "https://docs.sonsofphp.com"
+        "docs": "https://docs.sonsofphp.com/contracts/filesystem"
     },
     "autoload": {
         "psr-4": {
